create a WorkOrderUpdater class
